refactor(common): reorganize model classes for results into new package
