+++ xoai-common/src/main/java/io/gdcc/xoai/model/oaipmh/Granularity.java
@@ -8,8 +8,6 @@
 
 package io.gdcc.xoai.model.oaipmh;
 
-import java.time.format.DateTimeFormatter;
-
 /**
  * @author Development @ Lyncode
  * @version 3.1.0
