+++ xoai-common/src/main/java/io/gdcc/xoai/model/oaipmh/results/package-info.java
@@ -1,0 +1,1 @@
+package io.gdcc.xoai.model.oaipmh.results;
